more fixes

+++ src/training/train.py
@@ -31,10 +31,10 @@ def get_loss(model, images, texts, loss_img, loss_txt, args):
 
         # We gather tensors from all gpus to get more negatives to contrast with.
         gathered_image_features = [
-            torch.zeros_like(image_features, dtype=torch.half) for _ in range(world_size)
+            torch.zeros_like(image_features) for _ in range(world_size)
         ]
         gathered_text_features = [
-            torch.zeros_like(text_features, dtype=torch.half) for _ in range(world_size)
+            torch.zeros_like(text_features) for _ in range(world_size)
         ]
         dist.all_gather(gathered_image_features, image_features)  # comment this to measure without allgather
         dist.all_gather(gathered_text_features, text_features)  # comment this to measure without allgather
